Ajout structure if addition

+++ Calcul mental/calcul.c
@@ -1,16 +1,16 @@
-float addition(float nombrea, float nombreb)
+int addition(int nombrea, int nombreb)
 {
 	return nombrea + nombreb;
 }
-float multiplication(float nombrea, float nombreb)
+int multiplication(float nombrea, float nombreb)
 {
 	return nombrea * nombreb;
 }
 float division(float nombrea, float nombreb)
 {
-	return nombrea / nombreb;
+	return nombrea/nombreb;
 }
-float soustraction(float nombrea, float nombreb)
+int soustraction(float nombrea, float nombreb)
 {
 	return nombrea - nombreb;
 }
